% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.coxph.R
\name{spline.coxph}
\alias{spline.coxph}
\title{Fitting a parametric proportional hazard model with M-spline function}
\usage{
spline.coxph(
  t.event,
  event,
  Z,
  xi1 = min(t.event),
  xi3 = max(t.event),
  shape = "bathtub1",
  p0 = rep(0, 1 + p)
)
}
\arguments{
\item{t.event}{observed survival time}

\item{event}{event Indicator (0 or 1)}

\item{Z}{covariate vector}

\item{xi1}{minimum survival time}

\item{xi3}{maximum survival time}

\item{shape}{select shape of baseline hazard}

\item{p0}{initial value of parameters}
}
\value{
results of estimation
}
\description{
\code{spline.coxph} fitting a parametric proportional hazard model with M-spline function
}
